JAXB Upgrade
git-svn-id: https://hermesjms.svn.sourceforge.net/svnroot/hermesjms/trunk/HermesJMS@32 6ec42ede-3d11-0410-a3df-d4e4f9070504

+++ src/java/hermes/browser/dialog/GeneralConfigPanel.java
@@ -340,7 +340,7 @@ public class GeneralConfigPanel extends JPanel
                   }
                   else
                   {
-                     wConfig.setDefaultAgeAlertThreshold(((Integer) watchAgeAlertProperty.getValue()).intValue());
+                     wConfig.setDefaultAgeAlertThreshold(((Long) watchAgeAlertProperty.getValue()).longValue());
                   }
                }
 
@@ -357,7 +357,7 @@ public class GeneralConfigPanel extends JPanel
                   }
                   else
                   {
-                     wConfig.setUpdateFrequency(((Integer) watchFrequenceProperty.getValue()).intValue());
+                     wConfig.setUpdateFrequency(((Long) watchFrequenceProperty.getValue()).longValue());
                   }
                }
